Added checks for username and email, login service accepts user type in json as well
