+++ src/main/java/com/ubar/dao/UserDAO.java
@@ -10,11 +10,15 @@ import javax.transaction.Transactional;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.ubar.user.service.User;
+import com.ubar.model.User;
 
 @Transactional
 public interface UserDAO extends CrudRepository<User, Long>{
 	
-	public Optional<User> findByUsernameAndPassword(String username, String password);
+	public Optional<User> findByUsernameAndPasswordAndType(String username, String password, String type);
+	
+	public Optional<User> findByUsername(String username);
+	
+	public Optional<User> findByEmail(String email);
 
 }
