+++ src/main/java/com/ubar/user/dto/LoginResponse.java
@@ -10,6 +10,10 @@ public class LoginResponse {
 	
 	private String username, type;
 	
+	public LoginResponse() {
+		super();
+	}
+	
 	public LoginResponse(long id, String username, String type) {
 		super();
 		this.id = id;
